% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrlmgraphbench-package.R
\docType{package}
\name{rrlmgraphbench-package}
\alias{rrlmgraphbench}
\alias{rrlmgraphbench-package}
\title{rrlmgraphbench: Benchmark Suite for rrlmgraph}
\description{
Provides fixture projects and ground-truth datasets for benchmarking the rrlmgraph package. Fixtures range from a minimal data-science script project to a medium-sized Shiny application and a small R package, covering the full spectrum of project types supported by rrlmgraph.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidrsch/rrlmgraph-bench}
  \item Report bugs at \url{https://github.com/davidrsch/rrlmgraph-bench/issues}
}

}
\author{
\strong{Maintainer}: David Díaz \email{davidrsch@gmail.com} (\href{https://orcid.org/0000-0002-0927-9795}{ORCID}) [copyright holder]

}
\keyword{internal}
